third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/get-from-iframe-expected.txt

CONSOLE ERROR: line 4: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?notifyDone=1&q=%3Cscript%3Ealert%28String.fromCharCode%280x58%2C0x53%2C0x53%29%29%3C%2Fscript%3E' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
